test: add email e2e and isolate test-mode endpoints (#37)
