+++ .github/workflows/test.yml
@@ -154,5 +154,6 @@ jobs:
           DOTBOT_SERVER_URL: ${{ secrets.DOTBOT_SERVER_URL }}
           DOTBOT_API_KEY: ${{ secrets.DOTBOT_API_KEY }}
           DOTBOT_TEAMS_RECIPIENT: ${{ secrets.DOTBOT_TEAMS_RECIPIENT }}
+          DOTBOT_EMAIL_RECIPIENT: ${{ secrets.DOTBOT_EMAIL_RECIPIENT }}
         run: |
           pwsh -NoProfile -ExecutionPolicy Bypass -File tests/Run-Tests.ps1 -Layer 4
